sqlite_to_postgres/requirements.txt
psycopg2-binary==2.9.3
python-dotenv==0.19.2
flake8==4.0
python-dateutil==2.8.2
pytest==7.0.1
